Reorganize extra stage code into multiple files
