+++ src/stages/stagex/meson.build
@@ -1,0 +1,9 @@
+
+stagex_src = files(
+    'background_anim.c',
+    'draw.c',
+    'spells/infinity_net.c',
+    'stagex.c',
+    'timeline.c',
+    'yumemi.c',
+)
