+++ src/stages/stagex/timeline.c
@@ -1,0 +1,119 @@
+/*
+ * This software is licensed under the terms of the MIT License.
+ * See COPYING for further information.
+ * ---
+ * Copyright (c) 2011-2024, Lukas Weber <laochailan@web.de>.
+ * Copyright (c) 2012-2024, Andrei Alexeyev <akari@taisei-project.org>.
+ */
+
+#include "timeline.h"   // IWYU pragma: keep
+#include "stagex.h"
+#include "yumemi.h"
+
+#include "stages/common_imports.h"
+
+TASK(glider_bullet, {
+	cmplx pos; real dir; real spacing; int interval;
+}) {
+	const int nproj = 5;
+	const int nstep = 4;
+	BoxedProjectile projs[nproj];
+
+	const cmplx positions[][5] = {
+		{1+I, -1, 1, -I, 1-I},
+		{2, I, 1, -I, 1-I},
+		{2, 1+I, 2-I, -I, 1-I},
+		{2, 0, 2-I, 1-I, 1-2*I},
+	};
+
+	cmplx trans = cdir(ARGS.dir+1*M_PI/4)*ARGS.spacing;
+
+	for(int i = 0; i < nproj; i++) {
+		projs[i] = ENT_BOX(PROJECTILE(
+			.pos = ARGS.pos+positions[0][i]*trans,
+			.proto = pp_ball,
+			.color = RGBA(0,0,1,1),
+		));
+	}
+
+	for(int step = 0;; step++) {
+		int cur_step = step%nstep;
+		int next_step = (step+1)%nstep;
+
+		int dead_count = 0;
+		for(int i = 0; i < nproj; i++) {
+			Projectile *p = ENT_UNBOX(projs[i]);
+			if(p == NULL) {
+				dead_count++;
+			} else {
+				p->move.retention = 1;
+				p->move.velocity = -(positions[cur_step][i]-(1-I)*(cur_step==3)-positions[next_step][i])*trans/ARGS.interval;
+			}
+		}
+		if(dead_count == nproj) {
+			return;
+		}
+		WAIT(ARGS.interval);
+	}
+}
+
+TASK(glider_fairy, {
+	real hp; cmplx pos; cmplx dir;
+}) {
+	Enemy *e = TASK_BIND(espawn_big_fairy(VIEWPORT_W/2-10*I, ITEMS(
+		.power = 3,
+		.points = 5,
+	)));
+
+	YIELD;
+
+	for(int i = 0; i < 80; ++i) {
+		e->pos += cnormalize(ARGS.pos-e->pos)*2;
+		YIELD;
+	}
+
+	for(int i = 0; i < 3; i++) {
+		real aim = carg(global.plr.pos - e->pos);
+		INVOKE_TASK(glider_bullet, e->pos, aim-0.7, 20, 6);
+		INVOKE_TASK(glider_bullet, e->pos, aim, 25, 3);
+		INVOKE_TASK(glider_bullet, e->pos, aim+0.7, 20, 6);
+		WAIT(80-20*i);
+	}
+
+	for(;;) {
+		e->pos += 2*(creal(e->pos)-VIEWPORT_W/2 > 0)-1;
+		YIELD;
+	}
+}
+
+DEFINE_EXTERN_TASK(stagex_timeline) {
+	YIELD;
+
+	// goto enemies;
+
+	WAIT(3900);
+	STAGE_BOOKMARK(boss);
+
+	global.boss = stagex_spawn_yumemi(VIEWPORT_W/2 + 180*I);
+
+	boss_add_attack_task(global.boss, AT_SurvivalSpell, "Obliteration “Infinity Network”",
+		90, 80000, TASK_INDIRECT(BossAttack, stagex_spell_infinity_network), NULL);
+
+	STALL;
+
+	PlayerMode *pm = global.plr.mode;
+	StageExPreBossDialogEvents *e;
+	INVOKE_TASK_INDIRECT(StageExPreBossDialog, pm->dialog->StageExPreBoss, &e);
+	INVOKE_TASK_WHEN(&e->music_changes, common_start_bgm, "stagexboss");
+	WAIT_EVENT(&global.dialog->events.fadeout_began);
+
+	boss_engage(global.boss);
+
+	WAIT_EVENT(&global.boss->events.defeated);
+
+enemies:
+	for(int i = 0;;i++) {
+		INVOKE_TASK_DELAYED(60, glider_fairy, 2000, CMPLX(VIEWPORT_W*(i&1), VIEWPORT_H*0.5), 3*I);
+		WAIT(50+100*(i&1));
+	}
+}
